Fix nightly test failure March 20, 2025 [#465] (#467)
Fix "stat: path should be string, bytes, os.PathLike or integer, not
NoneType" error. Check all mnist ops execute on PR.

### Ticket
#465 

### Problem description
Issue arose from 
`if os.path.isfile(inputs_file_path):`
when inputs_file_path is None, this gives an "stat: path should be
string, bytes, os.PathLike or integer, not NoneType" error.
This was replaced with
`if inputs_file_path and os.path.isfile(inputs_file_path):`

+ added a workflow job that checks all mnist ops run on PR.

+++ tt_torch/dynamo/executor.py
@@ -451,7 +451,7 @@ class OpByOpExecutor(Executor):
 
         inputs_size = get_inputs_size(inputs)
 
-        large_input = inputs_size > self.op_memory_limit
+        large_input = inputs_size >= self.op_memory_limit
 
         obj = {
             "binary": binary,
@@ -512,7 +512,7 @@ class OpByOpExecutor(Executor):
                 self.execute_process = None
                 break
 
-        if os.path.isfile(inputs_file_path):
+        if inputs_file_path and os.path.isfile(inputs_file_path):
             try:
                 os.remove(inputs_file_path)
             except OSError:
